refactor: 게임 실행 시 입력받는 명령어에 대한 메소드를 enum 클래스로 추출
