+++ src/main/java/chess/ConsoleUIChessApplication.java
@@ -1,15 +1,15 @@
 package chess;
 
+import chess.controller.Command;
+import chess.controller.GameController;
 import chess.view.InputView;
 import chess.view.OutputView;
 
-public class Application {
-    public static final String START = "start";
-
+public class ConsoleUIChessApplication {
     public static void main(String[] args) {
         OutputView.printStartMSG();
         String startMsg = InputView.inputCommand();
-        if (START.equals(startMsg)) {
+        if (Command.isStart(startMsg)) {
             GameController.start();
         }
     }
